+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -89,6 +89,8 @@ public class Settings extends Config {
 
 	public static boolean enableTownyRestrictions = false;
 
+	public static boolean checkSpawnLocationInteractionResult = false;
+
 	public static boolean disableInventoryVerification = false;
 
 	/*